+++ ProjP3.Application/DTOs/Response/ProfessorDTO.cs
@@ -5,9 +5,9 @@ namespace ProjP3.Application.DTOs.Response;
 
 public class ProfessorDTO
 {
-    public ulong IdProfessor { get; set; }
+    public int IdProfessor { get; set; }
 
-    public ulong IdTitulo { get; set; }
+    public int IdTitulo { get; set; }
 
     public string TxNome { get; set; } = null!;
