+++ ProjP3.Application/InterfaceServices/IInstituicaoService.cs
@@ -12,14 +12,14 @@ namespace ProjP3.Application.InterfaceServices
     public interface IInstituicaoService
     {
         Task<Result<List<InstituicaoDTO>>> GetAllAsync();
-        Task<Result<InstituicaoDTO>> GetByIdAsync(ulong id);
+        Task<Result<InstituicaoDTO>> GetByIdAsync(int id);
         Task<Result<InstituicaoDTO>> AddAsync(InstituicaoCreateDTO instituicao);
         Task<Result<InstituicaoDTO>> UpdateAsync(InstituicaoUpdateDTO instituicao);
-        Task<Result<bool>> DeleteAsync(ulong id);
+        Task<Result<bool>> DeleteAsync(int id);
         Task<Result<List<InstituicaoDTO>>> GetInstituicaoBySiglaAsync(string sigla);
         Task<Result<List<InstituicaoDTO>>> GetInstituicaoByDescricaoAsync(string descricao);
-        Task<Result<List<InstituicaoDTO>>> GetInstituicoesByCursoAsync(ulong idCurso);
-        //Task<Result<InstituicaoDTO>> AdicionarCursoAInstituicaoAsync(ulong idInstituicao, ulong idCurso);
-        //Task<Result<InstituicaoDTO>> RemoverCursoDeInstituicaoAsync(ulong idInstituicao, ulong idCurso);
+        Task<Result<List<InstituicaoDTO>>> GetInstituicoesByCursoAsync(int idCurso);
+        //Task<Result<InstituicaoDTO>> AdicionarCursoAInstituicaoAsync(int idInstituicao, int idCurso);
+        //Task<Result<InstituicaoDTO>> RemoverCursoDeInstituicaoAsync(int idInstituicao, int idCurso);
     }
 }
