+++ ProjP3.Application/InterfaceServices/ITipoCursoService.cs
@@ -12,10 +12,10 @@ namespace ProjP3.Application.InterfaceServices
     public interface ITipoCursoService
     {
         Task<Result<List<TipoCursoDTO>>> GetAllAsync();
-        Task<Result<TipoCursoDTO>> GetByIdAsync(ulong id);
+        Task<Result<TipoCursoDTO>> GetByIdAsync(int id);
         Task<Result<TipoCursoDTO>> AddAsync(TipoCursoCreateDTO tipoCursoDto);
         Task<Result<TipoCursoDTO>> UpdateAsync(TipoCursoUpdateDTO tipoCursoDto);
-        Task<Result<bool>> DeleteAsync(ulong id);
+        Task<Result<bool>> DeleteAsync(int id);
         Task<Result<List<TipoCursoDTO>>> GetTipoCursoByDescricaoAsync(string descricao);
     }
 }
